Move UI to UI subpackage
