+++ features/auth/src/main/kotlin/com/boswelja/truemanager/auth/DI.kt
@@ -1,5 +1,6 @@
 package com.boswelja.truemanager.auth
 
+import com.boswelja.truemanager.auth.ui.AuthViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
